+++ render/src/engine.rs
@@ -1,7 +1,7 @@
 use super::page::Page;
 use flume::{Receiver, Sender};
 use gfx::Bitmap;
-use loader::{ResourceLoader, RESOURCE_LOADER};
+use loader::resource_loader::{ResourceLoader, InProcessResourceLoader};
 use shared::primitive::Size;
 use url::Url;
 
@@ -17,14 +17,14 @@ pub enum OutputEvent {
 
 pub struct RenderEngine<'a> {
     page: Page<'a>,
-    res_loader: ResourceLoader,
+    resource_loader: InProcessResourceLoader,
 }
 
 impl<'a> RenderEngine<'a> {
     pub async fn new(viewport: Size) -> RenderEngine<'a> {
         let page = Page::new(viewport).await;
-        let res_loader = ResourceLoader::init();
-        Self { page, res_loader }
+        let resource_loader = InProcessResourceLoader::new();
+        Self { page, resource_loader }
     }
 
     pub async fn run(
@@ -49,11 +49,7 @@ impl<'a> RenderEngine<'a> {
                 self.emit_new_frame(event_emitter)?;
             }
             InputEvent::LoadHTML { html, base_url } => {
-                RESOURCE_LOADER
-                    .scope(self.res_loader.clone(), async {
-                        self.page.load_html(html, base_url).await;
-                    })
-                    .await;
+                self.page.load_html(html, base_url).await;
                 self.emit_new_frame(event_emitter)?;
                 self.emit_new_title(event_emitter)?;
             }
